Player can pickup screwdriver and unlock power panel
The player can now pickup the screwdriver object (this functionality still needs to be extended to other objects - will add this later).

The player can then 'unlock' the 'power panel' iff they have picked up the screwdriver.

Currently only the screwdriver object is being tracked in the inventory, but I will extend this to more objects.