completed room 5, needs photon integration
Room 5 completed, missing exit doors+ photon integration

At least 2 players are needed to complete this puzzle. Players spawn in one of two rooms with coloured squares on the floor, coloured boxes and coloured buttons. Players must push the coloured boxes onto the correct coloured squares such that the rooms are mirrored; there is a window to help players with this. Each room has 3 coloured squares on the ceiling each corresponding to a correct box position in that room (floating yellow square is the correct position for the yellow box ect.). Once the boxes are in the correct positions the doors should spawn and the puzzle should be solved. If the boxes are moves once this happens the exit doors should stay there. The coloured buttons respawn their corresponding coloured boxes, incase a box leaves the room, or the players want to reset that box for any reason.