Additions to Room2 Floor1
Added Buttons for players to move box in x z plane
Once box has moved in a direction, the direction can't be changed until it collides with an orange object and stops moving.
Box position can be reset with button
All things that need to be synced are bundled into functions for easy photon integration.
Need to add actual path to move box through, goal for box to reach, split rooms, duplicate button, integrate with photon.