Integrated Room 1 Floor 2 puzzleroom with Photon
Fully Integrated the puzzlerrom with photon, only thing that needs to be added are scripts for the exit doors. There's a temporary door to the puzzleroom located in the first puzzleroom on the left in the second floor, message me if you can't find it (This will probably be where the door will be in the final version.

Puzzle Description:
You must move the brown box using the movement buttons through the gap in the wall in the top right hand corner of the room. The area players can move in is split into two halves, one half has buttons to move the box up and down and the other left and right. Players need to communicate to move the box in the correct direction. Players are able to move the box only when they move it for the first time or if the box has collided with an ORANGE object. If the box collides with a GREEN object the box will have to be reset. Players can reset the box by pressing the buttons on the wall, this can be done any time. If you want to know the solution if you are testing the room message me and I'll send it to you.