Added Room Managers
Added Room managers for each floor which move the walls blocking stairs/ exit when all puzzlerooms in that floor have been completed.

Each Manager has a function PuzzleComplete which can be called by a puzzleroom when its puzzle is completed.

The number of puzzles variable has been set to 0 for testing purposes, so the walls are already moved at the start of the game