Textured moving path puzzleroom. Puzzle description in commit desc,
<2 players.
They have to get to the top of the room by climbing stairs and crossing paths
Every few seconds one player will see one of the paths' collider/renderer be enabled, showing the correct path to travel on. The other player does not see the renderer being enabled/disabled, so the player who sees has to communicate with other player.
Players roles are then reversed.
For the final path, there are 2 sections. Players are only able to see the state of one of the sections.
One player presses button at the end to spawn exit doos & complete puzzle